Added Simple Auto Mod
Not a fully feature auto mod, rather a temporary fix to keep the server safe from yes, you guessed it! Mass pingers aka Thee Alt God. 

Although the sole purpose of this code is to mute mass pingers, it still has some flaws. The major one being not able to detect pings in multiple message. It will mute someone if and only if the amount of pings in a single message reaches the 5 pings per message limit. However, if anyone pings at most 4 times in a message and keeps doing it in several messages, the bot won't do a thing.